Split all API methods into extensions
